Force dtype to float for balance
If dtype is not set to dfloat, pandas produce warnings when merging the balance in portfolio.py. Another alternative would to rewrite the integer in the solution as explicit floats (1 -> 1.0).

+++ apis.py
@@ -50,7 +50,7 @@ class Api(ABC):
     @staticmethod
     def _balance_to_dataframe(balance: dict, api_name: str) -> pd.DataFrame:
         '''Convert dictionary balance into a pandas dataframe.'''
-        df = pd.DataFrame.from_dict(balance, orient = 'index')
+        df = pd.DataFrame.from_dict(balance, orient = 'index', dtype='float')
         df.reset_index(inplace = True)
         df.columns = Api.col_names
         df['API'] = api_name
